added custom not found exception

+++ Gamestore.Application/Features/Categories/Commands/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Gamestore.Application.Contracts.Persistance;
+using Gamestore.Application.Exceptions;
 using MediatR;
 
 namespace Gamestore.Application.Features.Categories.Commands.DeleteCategory;
@@ -15,6 +16,12 @@ public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryComman
     public async Task<Unit> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
     {
         var categoryToDelete = await _categoryRepository.GetByIdAsync(request.Id);
+
+        if (categoryToDelete == null)
+        {
+            throw new NotFoundException(nameof(categoryToDelete), request.Id);
+        }
+
         await _categoryRepository.DeleteAsync(categoryToDelete);
 
         return Unit.Value;
