+++ Gamestore.Application/Features/Comments/Commands/DeleteComment/DeleteCommentCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Gamestore.Application.Contracts.Persistance;
+using Gamestore.Application.Exceptions;
 using MediatR;
 
 namespace Gamestore.Application.Features.Comments.Commands.DeleteComment;
@@ -15,6 +16,12 @@ public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand,
     public async Task<Unit> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
     {
         var commentToDelete = await _commentRepository.GetByIdAsync(request.Id);
+
+        if (commentToDelete == null)
+        {
+            throw new NotFoundException(nameof(commentToDelete), request.Id);
+        }
+
         await _commentRepository.DeleteAsync(commentToDelete);
 
         return Unit.Value;
