+++ Gamestore.Application/Features/Games/Commands/DeleteGame/DeleteGameCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Gamestore.Application.Contracts.Persistance;
+using Gamestore.Application.Exceptions;
 using MediatR;
 
 namespace Gamestore.Application.Features.Games.Commands.DeleteGame;
@@ -15,6 +16,12 @@ public class DeleteGameCommandHandler : IRequestHandler<DeleteGameCommand, Unit>
     public async Task<Unit> Handle(DeleteGameCommand request, CancellationToken cancellationToken)
     {
         var gameToDelete = await _gameRepository.GetByIdAsync(request.Id);
+
+        if (gameToDelete == null)
+        {
+            throw new NotFoundException(nameof(gameToDelete), request.Id);
+        }
+
         await _gameRepository.DeleteAsync(gameToDelete);
 
         return Unit.Value;
