+++ Gamestore.Application/Features/Genres/Commands/DeleteGenre/DeleteGenreCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Gamestore.Application.Contracts.Persistance;
+using Gamestore.Application.Exceptions;
 using MediatR;
 
 namespace Gamestore.Application.Features.Genres.Commands.DeleteGenre;
@@ -15,6 +16,12 @@ public class DeleteGenreCommandHandler : IRequestHandler<DeleteGenreCommand, Uni
     public async Task<Unit> Handle(DeleteGenreCommand request, CancellationToken cancellationToken)
     {
         var genreToDelete = await _genreRepository.GetByIdAsync(request.Id);
+
+        if (genreToDelete == null)
+        {
+            throw new NotFoundException(nameof(genreToDelete), request.Id);
+        }
+
         await _genreRepository.DeleteAsync(genreToDelete);
 
         return Unit.Value;
