+++ Gamestore.Application/Features/Orders/Commands/DeleteOrder/DeleteOrderCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Gamestore.Application.Contracts.Persistance;
+using Gamestore.Application.Exceptions;
 using MediatR;
 
 namespace Gamestore.Application.Features.Orders.Commands.DeleteOrder;
@@ -15,6 +16,12 @@ public class DeleteOrderCommandHandler : IRequestHandler<DeleteOrderCommand, Uni
     public async Task<Unit> Handle(DeleteOrderCommand request, CancellationToken cancellationToken)
     {
         var orderToDelete = await _orderRepository.GetByIdAsync(request.Id);
+
+        if (orderToDelete == null)
+        {
+            throw new NotFoundException(nameof(orderToDelete), request.Id);
+        }
+
         await _orderRepository.DeleteAsync(orderToDelete);
 
         return Unit.Value;
