+++ Gamestore.Application/Features/OrdersDetails/Commands/DeleteOrderDetails/DeleteOrderDetailsCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Gamestore.Application.Contracts.Persistance;
+using Gamestore.Application.Exceptions;
 using MediatR;
 
 namespace Gamestore.Application.Features.OrdersDetails.Commands.DeleteOrderDetails;
@@ -16,6 +17,12 @@ public class DeleteOrderDetailsCommandHandler : IRequestHandler<DeleteOrderDetai
     public async Task<Unit> Handle(DeleteOrderDetailsCommand request, CancellationToken cancellationToken)
     {
         var orderDetailsToDelete = await _orderDetailsRepository.GetByIdAsync(request.Id);
+
+        if (orderDetailsToDelete == null)
+        {
+            throw new NotFoundException(nameof(orderDetailsToDelete), request.Id);
+        }
+
         await _orderDetailsRepository.DeleteAsync(orderDetailsToDelete);
 
         return Unit.Value;
