+++ Gamestore.Application/Features/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Gamestore.Application.Contracts.Persistance;
+using Gamestore.Application.Exceptions;
 using MediatR;
 
 namespace Gamestore.Application.Features.Users.Commands.DeleteUser;
@@ -15,6 +16,12 @@ public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, Unit>
     public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
         var userToDelete = await _userRepository.GetByIdAsync(request.Id);
+
+        if (userToDelete == null)
+        {
+            throw new NotFoundException(nameof(userToDelete), request.Id);
+        }
+
         await _userRepository.DeleteAsync(userToDelete);
 
         return Unit.Value;
